challenge results API initial hook
