+++ packages/front-end/src/views/challenge/hooks/use-challenge-api.ts
@@ -1,14 +1,10 @@
-import { useAlerts } from "@tiernebre/kecleon";
 import { useState, useMemo, useCallback } from "react";
 import {
   Challenge,
-  ChallengeResult,
-  HttpChallengeParticipantService,
   HttpChallengeService,
   HttpClient,
   SessionPayload,
 } from "../../../api";
-import { ChallengeParticipantUpdateRequest } from "../../../api/challenge/ChallengeParticipantUpdateRequest";
 
 export type ChallengeHookParameters = {
   challengeId: number;
@@ -18,12 +14,8 @@ export type ChallengeHookParameters = {
 
 export type ChallengeHookReturnValue = {
   challenge: Challenge | undefined;
-  results: ChallengeResult[] | undefined;
-  existingResultForUser: ChallengeResult | undefined;
-  userIsInChallenge: boolean;
   userOwnsChallenge: boolean;
   fetchChallenge: () => Promise<void>;
-  submitResult: (request: ChallengeParticipantUpdateRequest) => Promise<void>;
 };
 
 export const useChallengeApi = ({
@@ -32,58 +24,21 @@ export const useChallengeApi = ({
   session,
 }: ChallengeHookParameters): ChallengeHookReturnValue => {
   const [challenge, setChallenge] = useState<Challenge>();
-  const [results, setResults] = useState<ChallengeResult[]>([]);
-  const { showAlert } = useAlerts();
 
   const challengeService = useMemo(
     () => new HttpChallengeService(httpClient),
     [httpClient]
   );
-  const challengeParticipantService = useMemo(
-    () => new HttpChallengeParticipantService(httpClient),
-    [httpClient]
-  );
 
-  const existingResultForUser = results.find(
-    (result) => result.participantId === session.userId
-  );
-  const userIsInChallenge = !!existingResultForUser;
   const userOwnsChallenge = challenge?.creatorId === session.userId;
 
   const fetchChallenge = useCallback(async () => {
     setChallenge(await challengeService.getOneById(challengeId));
-    setResults(await challengeService.getResultsForChallenge(challengeId));
   }, [challengeService, challengeId]);
-
-  const submitResult = useCallback(
-    async (request: ChallengeParticipantUpdateRequest) => {
-      if (existingResultForUser) {
-        await challengeParticipantService.updateOne(
-          existingResultForUser.resultId,
-          request
-        );
-        await fetchChallenge();
-        showAlert({
-          message: "Challenge Submission Successfully Submitted",
-          color: "success",
-        });
-      }
-    },
-    [
-      challengeParticipantService,
-      existingResultForUser,
-      showAlert,
-      fetchChallenge,
-    ]
-  );
 
   return {
     challenge,
-    results,
-    existingResultForUser,
     fetchChallenge,
-    userIsInChallenge,
     userOwnsChallenge,
-    submitResult,
   };
 };
